requirements.txt
aiohttp==3.7.4
nest_asyncio==1.5.1
aiogram==2.14.3
protobuf==3.17.3
python-dotenv==0.19.0
google-cloud-bigquery==2.26.0